# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Dosen",
        "description": "Dokumentasi API untuk manajemen data dosen",
        "version": "1.0"
    },
    "paths": {
...
        "/api/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Menampilkan daftar mahasiswa",
                "operationId": "5cd79af38ea07283ea41db31685fdadb",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "Filter berdasarkan nama",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Menambahkan mahasiswa",
                "operationId": "2b325793b6d9bd67deb73fd27727cae1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nim",
                                    "nama",
                                    "jenis_kelamin",
                                    "alamat",
                                    "tanggal_lahir",
                                    "program_studi",
                                    "angkatan",
                                    "email",
                                    "status",
                                    "agama"
                                ],
                                "properties": {
                                    "nim": {
                                        "type": "string"
                                    },
                                    "nama": {
                                        "type": "string"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "program_studi": {
                                        "type": "string"
                                    },
                                    "angkatan": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "agama": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                }
            }
        },
        "/api/mahasiswa/nim/{nim}": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Menampilkan data mahasiswa berdasarkan NIM",
                "operationId": "2afe8e056b3dd889df6e99e4cef7b8e1",
                "parameters": [
                    {
                        "name": "nim",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mengupdate data mahasiswa berdasarkan NIM",
                "operationId": "60fbdf2e61eb30ddee4bc31f2a5e5f17",
                "parameters": [
                    {
                        "name": "nim",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Mahasiswa"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Menghapus mahasiswa berdasarkan NIM",
                "operationId": "1751ca8f2e951ce71a52e75e2faf69b6",
                "parameters": [
                    {
                        "name": "nim",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
        "/api/mahasiswa/{id}": {
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Mengupdate data mahasiswa berdasarkan ID",
                "operationId": "09000b1b4250504e890bf4b52ce4e195",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Mahasiswa"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Mahasiswa": {
                "required": [
                    "nim",
                    "nama",
                    "jenis_kelamin",
                    "alamat",
                    "tanggal_lahir",
                    "program_studi",
                    "angkatan",
                    "email",
                    "status",
                    "agama"
                ],
                "properties": {
                    "nim": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "jenis_kelamin": {
                        "type": "string",
                        "example": "L"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka 123"
                    },
                    "tanggal_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01"
                    },
                    "program_studi": {
                        "type": "string",
                        "example": "Teknik Informatika"
                    },
                    "angkatan": {
                        "type": "integer",
                        "example": 2020
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif"
                    },
                    "agama": {
                        "type": "string",
                        "example": "Islam"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}